SERVER-24 chore: 패키지 구조를 domain 하위의 applemusic으로 수정
