+++ src/main/java/org/sopt/confeti/domain/applemusic/artist/infra/repository/ArtistRepository.java
@@ -1,6 +1,6 @@
-package org.sopt.confeti.domain.artist.infra.repository;
+package org.sopt.confeti.domain.applemusic.artist.infra.repository;
 
-import org.sopt.confeti.domain.artist.Artist;
+import org.sopt.confeti.domain.applemusic.artist.Artist;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ArtistRepository extends JpaRepository<Artist, String> {
