+++ src/main/java/org/sopt/confeti/domain/applemusic/artist_song/infra/repository/ArtistSongRepository.java
@@ -1,6 +1,6 @@
-package org.sopt.confeti.domain.artist_song.infra.repository;
+package org.sopt.confeti.domain.applemusic.artist_song.infra.repository;
 
-import org.sopt.confeti.domain.artist_song.ArtistSong;
+import org.sopt.confeti.domain.applemusic.artist_song.ArtistSong;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ArtistSongRepository extends JpaRepository<ArtistSong, Long> {
