+++ src/main/java/org/sopt/confeti/domain/applemusic/song/infra/repository/SongRepository.java
@@ -1,6 +1,6 @@
-package org.sopt.confeti.domain.song.infra.repository;
+package org.sopt.confeti.domain.applemusic.song.infra.repository;
 
-import org.sopt.confeti.domain.song.Song;
+import org.sopt.confeti.domain.applemusic.song.Song;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SongRepository extends JpaRepository<Song, String> {
