+++ src/main/java/org/sopt/confeti/domain/applemusic/top_artist/infra/repository/TopArtistRepository.java
@@ -1,6 +1,6 @@
-package org.sopt.confeti.domain.top_artist.infra.repository;
+package org.sopt.confeti.domain.applemusic.top_artist.infra.repository;
 
-import org.sopt.confeti.domain.top_artist.TopArtist;
+import org.sopt.confeti.domain.applemusic.top_artist.TopArtist;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TopArtistRepository extends JpaRepository<TopArtist, Long> {
